// src/types/style.ts
export type Theme = {
  backgroundColor1: string;
  backgroundColor2: string;
  backgroundColor3: string;
  paragraphColor1: string;
  paragraphColor2: string;
  paragraphColor3: string;
  headingColor: string;
  color2: string;
  color3: string;
  borderColor: string;
  linkColor1: string;
  linkColor2: string;
};

export type GlobalStyleProps = {
  theme: Theme;
};